Feature: CSV With Dynamic Class Creation and Testing for multiplication

+++ src/multiplication.py
@@ -2,4 +2,4 @@ def multiplication(a, b):
     a = float(a)
     b = float(b)
     c = a * b
-    return int(c)
+    return float(c)
